ADDED digraphclass.py, graphclass.py     This finalizes the refactoring of the property graph data structure to an array-centric data structure.
