+++ arachne/client/arachne/__init__.py
@@ -1,3 +1,4 @@
-from arachne.graphclasses import *
-from arachne.methods import *
+from arachne.graphclass import *
+from arachne.digraphclass import *
 from arachne.propgraphclass import *
+from arachne.methods import *
