--- arachne/client/arachne/graphclasses.py
@@ -1,590 +1,0 @@
-"""Contains the three graph classes that users can build: `Graph` (for undirected graphs), `DiGraph` 
-(for directed graphs), and `PropGraph` (for property graphs that are inherently directed). `Graph` 
-and `DiGraph` support weights on edges, but `PropGraph` does not since it supports types and 
-properties on vertices and edges.
-"""
-
-from __future__ import annotations
-from typing import cast, Tuple, Union
-import time
-import arkouda as ak
-import numpy as np
-from arkouda.client import generic_msg
-from arkouda.pdarrayclass import pdarray, create_pdarray
-from arkouda.logger import getArkoudaLogger
-from arkouda.dtypes import (
-    int64 as akint,
-    resolve_scalar_dtype
-)
-
-__all__ = ["Graph", "DiGraph"]
-
-class Graph:
-    """Base class for undirected graphs. 
-
-    This is the double index graph data structure based graph representation. The graph data resides 
-    on the arkouda server. The user should not call this class directly; rather its instances are 
-    created by other arachne functions.
-
-    Graphs hold undirected edges. Self loops and multiple edges are not allowed.
-
-    Nodes currently are only allowed to be integers. They may contain optional key/value attributes.
-
-    Edges are represented as links between nodes. They may contain optinal key/value attributes. 
-
-    Parameters
-    ----------
-    None
-
-    Attributes
-    ----------
-    n_vertices : int
-        The number of vertices in the graph. 
-    n_edges : int
-        The number of edges in the graph.
-    directed : int
-        The graph is directed (True) or undirected (False).
-    weighted : int
-        The graph is weighted (True) or unweighted (False).
-    name : string
-        The name of the graph in the backend Chapel server. 
-    logger : ArkoudaLogger
-        Used for all logging operations.
-
-    See Also
-    --------
-    DiGraph
-        
-    Notes
-    -----
-    """
-
-    def __init__(self, *args) -> None:
-        """Initializes the Graph instance by setting all instance
-        attributes, some of which are derived from the array parameters.
-        
-        Parameters
-        ----------
-        n_vertices
-            Must be provided in args[0].
-        n_edges
-            Must be provided in args[1].
-        weighted
-            Must be provided in args[3].
-        name
-            Must be provided in args[4].
-        directed
-            Defaults to 0 since Graph is inherently undirected. If needed,
-            should be found in args[2].
-        
-            
-        Returns
-        -------
-        None
-        
-        Raises
-        ------
-        RuntimeError
-            Raised if there's an error from the server to create a graph.  
-        ValueError
-            Raised if not enough arguments are passed to generate the graph. 
-        """
-        try:
-            if len(args) < 5:
-                self.n_vertices = 0
-                self.n_edges = 0
-                self.weighted = None
-                self.directed = 0
-                self.name = None
-            else:
-                self.n_vertices = int(cast(int, args[0]))
-                self.n_edges = int(cast(int, args[1]))
-                self.weighted = int(cast(int, args[3]))
-                oriname = cast(str, args[4])
-                self.name = oriname.strip()
-                self.directed = 0
-        except Exception as error:
-            raise RuntimeError(error)
-
-        self.dtype = akint
-        self.logger = getArkoudaLogger(name=__class__.__name__)
-
-    def __len__(self) -> int:
-        """Returns the number of vertices in the graph. Use: 'len(G)'.
-
-        Returns
-        -------
-        n_vertices: int.
-            The number of vertices in the graph.
-        """
-        return self.n_vertices
-
-    def size(self) -> int:
-        """Returns the number of edges in the graph. Use: 'G.size()'.
-
-        Returns
-        -------
-        n_edges: int.
-            The number of edges in the graph.
-        """
-        return self.n_edges
-
-    def nodes(self) -> pdarray:
-        """Returns the nodes of the graph as a pdarray. Use: 'G.nodes()'.
-
-        Returns
-        -------
-        nodes: pdarray.
-            The array containing the vertex information of a graph.
-        """
-        cmd = "nodes"
-        args = {"GraphName" : self.name}
-        repMsg = generic_msg(cmd=cmd, args=args)
-        returned_vals = (cast(str, repMsg).split('+'))
-
-        return create_pdarray(returned_vals[0])
-
-    def edges(self) -> Tuple[pdarray, pdarray]:
-        """Returns a tuple of pdarrays src and dst Use: 'G.edges()'.
-
-        Returns
-        -------
-        (src, dst): tuple.
-            The arrays containing the edge information of a graph.
-        """
-        cmd = "edges"
-        args = {"GraphName" : self.name}
-        repMsg = generic_msg(cmd=cmd, args=args)
-        returned_vals = (cast(str, repMsg).split('+'))
-
-        return create_pdarray(returned_vals[0]), create_pdarray(returned_vals[1])
-
-    def degree(self) -> pdarray:
-        """Returns the degree view for the whole graph as a pdarray. The format is NOTE: self-loops 
-        count twice to the degree count due to the sum of degree theorem.
-
-        Returns
-        -------
-        degree: pdarray
-            The array containing the degree for each node.
-        """
-        src,dst = self.edges()
-        degree = ak.GroupBy(src, assume_sorted=True).count()[1]
-        self_loops = src == dst
-        degree[src[self_loops]] += 1
-
-        return degree
-
-    def add_edges_from(self, akarray_src:pdarray, akarray_dst:pdarray,
-                       akarray_weight:Union[None,pdarray] = None) -> None:
-        """
-        Populates the graph object with edges as defined by the akarrays. Uses an Arkouda-based
-        reading version. 
-
-        Returns
-        -------
-        None
-        """
-        cmd = "addEdgesFrom"
-
-        ### Edge dedupping and delooping.
-        # 1. Symmetrize the graph by concatenating the src to dst arrays and vice versa.
-        src = ak.concatenate([akarray_src, akarray_dst])
-        dst = ak.concatenate([akarray_dst, akarray_src])
-        num_original_edges = src.size
-
-        # 1a. Initialize and symmetrize the weights of each edge, if applicable.
-        wgt = ak.array([1.0])
-        if isinstance(akarray_weight,pdarray):
-            wgt = ak.concatenate([akarray_weight, akarray_weight])
-            self.weighted = 1
-
-        # 2. Sort the edges and remove duplicates.
-        gb_edges = ak.GroupBy([src, dst])
-        src = gb_edges.unique_keys[0]
-        dst = gb_edges.unique_keys[1]
-        if bool(self.weighted):
-            wgt = gb_edges.aggregate(wgt, "sum")[1]
-
-        # 3. Calculate the number of edges that were removed. Due to symmetrization there will be
-        #    an extra copy of self-loops. If the total number of edges removed is z after
-        #    symmetrization then x of these are an extra copy of the self-loop. If there are copies
-        #    of an edge either as u~v or v~u then there will be an extra copy yielded after
-        #    symmetrizing. We will refer to the number of dedupped edges as y. They yield a total of
-        #    2y to the total number of edges removed, z. Therefore, we are solving the basic
-        #    algebraic problem of 2y + x = z where x and z are known for the graph after performing
-        #    the dedupping in the step above.
-        self_loops = src == dst
-        self_loops = ak.value_counts(self_loops)[1]
-        if self_loops.size == 1:
-            self_loops = 0 # This is our x.
-        else:
-            self_loops = self_loops[1]  # This is our x.
-
-        num_edges_after_dedup = src.size
-        num_removed_edges = num_original_edges - num_edges_after_dedup # This is our z.
-        num_dupped_edges = (num_removed_edges - self_loops) / 2 # This is solving for y.
-        num_edges = akarray_src.size - num_dupped_edges
-        
-        ### Vertex remapping.
-        # 1. Extract the unique vertex names of the graph.
-        gb_src_vertices = ak.GroupBy(src)
-        gb_dst_vertices = ak.GroupBy(dst)
-        vmap = gb_src_vertices.unique_keys
-
-        # 2. Create evenly spaced array within the range of the size of unique keys and broadcast
-        #    the values of the new range to the original vertices.
-        new_vertex_range = ak.arange(gb_src_vertices.unique_keys.size)
-        src = gb_src_vertices.broadcast(new_vertex_range)
-        dst = gb_dst_vertices.broadcast(new_vertex_range)
-
-        ### Create vertex index arrays.
-        # 1. Build the neighbors of the adjacency lists for each vertex.
-        gb_vertices = ak.GroupBy(src, assume_sorted=True)
-        gb_src_neighbors = gb_vertices.count()[1]
-
-        # 2. Run a prefix (cumulative) sum on neis to get the starting indices for each vertex.
-        segs = ak.cumsum(gb_src_neighbors)
-        first_seg = ak.array([0])
-        segs = ak.concatenate([first_seg, segs])
-
-        # 3. Extract the number of vertices, edges, and weighted flag from the graph.
-        self.n_vertices = int(vmap.size)
-        self.n_edges = int(num_edges)
-
-        ### Store everything in a graph object in the Chapel server.
-        # 1. Store data into an Graph object in the Chapel server.
-        args = { "AkArraySrc" : src,
-                 "AkArrayDst" : dst,
-                 "AkArraySeg" : segs,
-                 "AkArrayWeight" : wgt,
-                 "AkArrayVmap" : vmap,
-                 "Directed": bool(self.directed),
-                 "Weighted" : bool(self.weighted),
-                 "NumVertices" : self.n_vertices,
-                 "NumEdges" : self.n_edges }
-
-        rep_msg = generic_msg(cmd=cmd, args=args)
-        oriname = rep_msg
-        self.name = oriname.strip()
-
-    def add_edges_from_compat(self, akarray_src:pdarray, akarray_dst:pdarray) -> None:
-        """
-        Populates the graph object with edges as defined by the akarrays. Uses an Arkouda-based
-        reading version, is here to provide compatibility with the original algorithmic 
-        implementations for triangle counting, triangle centrality, k-truss, and connected
-        components.
-
-        Returns
-        -------
-        None
-        """
-        cmd = "addEdgesFromCompat"
-        src, dst = akarray_src, akarray_dst
-        srcR, dstR = akarray_dst, akarray_src
-
-        ### Remove self-loops from both the regular arrays and the reversed arrays.
-        self_loops = src == dst
-        src = src[~self_loops]
-        dst = dst[~self_loops]
-        srcR = srcR[~self_loops]
-        dstR = dstR[~self_loops]
-
-        ### Remove dupllicated edges from the graph.
-        # 1. First, for the regular edges.
-        regular_edges_gb = ak.GroupBy([src, dst])
-        src = regular_edges_gb.unique_keys[0]
-        dst = regular_edges_gb.unique_keys[1]
-
-        # 2. Secondly, for the reversed edges.
-        reversed_edges_gb = ak.GroupBy([srcR, dstR])
-        srcR = reversed_edges_gb.unique_keys[0]
-        dstR = reversed_edges_gb.unique_keys[1]
-
-        ### Remap the vertex names to a one-up mapping.
-        # 1. Get the vertex mapping.
-        vertices_gb = ak.GroupBy(ak.concatenate([src,dst]))
-        vertices = vertices_gb.unique_keys
-
-        # 2. Concatenate all the arrays and broadcast new values to them.
-        new_vertex_range = ak.arange(vertices.size)
-        all_vertices = ak.concatenate([src,dst,srcR,dstR])
-        all_vertices_gb = ak.GroupBy(all_vertices)
-        vmap = all_vertices_gb.unique_keys
-        all_vertices = all_vertices_gb.broadcast(new_vertex_range)
-
-        # 3. Extract the high ranges for each array in the big concatenated GroupBy.
-        src_high = src.size
-        dst_high = src_high + src.size
-        srcR_high = dst_high + src.size
-        dstR_high = srcR_high + src.size
-
-        # 4. Extract the actual arrays with slicing.
-        src = all_vertices[0:src_high]
-        dst = all_vertices[src_high:dst_high]
-        srcR = all_vertices[dst_high:srcR_high]
-        dstR = all_vertices[srcR_high:dstR_high]
-
-        ### Create vertex index arrays.
-        # 1. Create full arrays of the size of the vertex set.
-        nei = ak.full(vmap.size, 0, int)
-        neiR = ak.full(vmap.size, 0, int)
-        start_i = ak.full(vmap.size, -1, int)
-        start_iR = ak.full(vmap.size, -1, int)
-
-        # 2. Extract the neighbor count by doing a count on the number of times each vertex appears
-        #    in src.
-        gb_src = ak.GroupBy(src)
-        gb_src_indices, gb_src_neighbors = gb_src.count()
-        nei[gb_src_indices] = gb_src_neighbors
-
-        # 2a. Same as 2 but for srcR.
-        gb_srcR = ak.GroupBy(srcR)
-        gb_srcR_indices, gb_srcR_neighbors = gb_srcR.count()
-        neiR[gb_srcR_indices] = gb_srcR_neighbors
-
-        # 3. Find where each vertex starts inside of src and srcR.
-        start_i = ak.find(new_vertex_range, src)
-        start_iR = ak.find(new_vertex_range, srcR)
-
-        # 4. Extract vertex and edge number information.
-        self.n_vertices = int(vmap.size)
-        self.n_edges = int(src.size)
-        self.weighted = 0
-
-        ### Store everything in a graph object in the Chapel server.
-        args = { "AkArraySrc" : src,
-                 "AkArrayDst" : dst,
-                 "AkArraySrcR" : srcR,
-                 "AkArrayDstR" : dstR,
-                 "AkArrayNei" : nei,
-                 "AkArrayNeiR" : neiR,
-                 "AkArrayStartIdx" : start_i,
-                 "AkArrayStartIdxR" : start_iR,
-                 "AkArrayVmap" : vmap,
-                 "Directed": bool(self.directed),
-                 "Weighted" : bool(self.weighted),
-                 "NumVertices" : self.n_vertices,
-                 "NumEdges" : self.n_edges }
-
-        rep_msg = generic_msg(cmd=cmd, args=args)
-        oriname = rep_msg
-        self.name = oriname.strip()
-
-class DiGraph(Graph):
-    """Base class for directed graphs. Inherits from Graph.
-
-    This is the double index graph data structure based graph representation. The graph data resides 
-    on the arkouda server. The user should not call this class directly; rather its instances are 
-    created by other arachne functions.
-
-    DiGraphs hold directed edges. Self loops and multiple edges are not allowed.
-
-    Nodes currently are only allowed to be integers. They may contain optional key/value attributes.
-
-    Edges are represented as directed links between nodes. They may contain optional key/value 
-    attributes. 
-
-    Parameters
-    ----------
-    None
-
-    Attributes
-    ----------
-    n_vertices : int
-        The number of vertices in the graph. 
-    n_edges : int
-        The number of edges in the graph.
-    directed : int
-        The graph is directed (True) or undirected (False).
-    weighted : int
-        The graph is weighted (True) or unweighted (False).
-    name : string
-        The name of the graph in the backend Chapel server. 
-    logger : ArkoudaLogger
-        Used for all logging operations.
-
-    See Also
-    --------
-    Graph
-        
-    Notes
-    -----
-    """
-
-    def __init__(self, *args) -> None:
-        """Initializes the Graph instance by setting all instance
-        attributes, some of which are derived from the array parameters.
-        
-        Parameters
-        ----------
-        n_vertices
-            Must be provided in args[0].
-        n_edges
-            Must be provided in args[1].
-        weighted
-            Must be provided in args[3].
-        name
-            Must be provided in args[4].
-        directed
-            Defaults to 1 since DiGraph is inherently directed. If needed,
-            should be found in args[2].
-         
-        Returns
-        -------
-        None
-        
-        Raises
-        ------
-        RuntimeError
-            Raised if there's an error from the server to create a graph.   
-        ValueError
-            Raised if not enough arguments are passed to generate the graph. 
-        """
-        try:
-            if len(args) < 5:
-                self.n_vertices = 0
-                self.n_edges = 0
-                self.weighted = None
-                self.directed = 1
-                self.name = None
-            else:
-                self.n_vertices = int (cast(int, args[0]))
-                self.n_edges = int(cast(int, args[1]))
-                self.weighted = int(cast(int, args[3]))
-                oriname = cast(str, args[4])
-                self.name = oriname.strip()
-                self.directed = 1
-        except Exception as e:
-            raise RuntimeError(e)
-
-        self.dtype = akint
-        self.logger = getArkoudaLogger(name=__class__.__name__)
-
-    def out_degree(self) -> pdarray:
-        """Returns the out degree view for the whole graph as a pdarray.
-
-        Returns
-        -------
-        out_degree: pdarray
-            The array containing the out_degrees for each node.
-        """
-        src = self.edges()[0]
-        
-        out_degree_keys,out_degree_count = ak.GroupBy(src, assume_sorted=True).count()
-        out_degree = ak.full(self.n_vertices, 0, dtype=ak.int64)
-        out_degree[out_degree_keys] = out_degree_count
-
-        return out_degree
-    
-    def in_degree(self) -> pdarray:
-        """Returns the out degree view for the whole graph as a pdarray.
-
-        Returns
-        -------
-        in_degree: pdarray
-            The array containing the in_degrees for each node.
-        """
-        dst = self.edges()[1]
-
-        in_degree_keys,in_degree_count = ak.GroupBy(dst).count()
-        in_degree = ak.full(self.n_vertices, 0, dtype=ak.int64)
-        in_degree[in_degree_keys] = in_degree_count
-
-        return in_degree
-
-    def add_edges_from(self, akarray_src:pdarray, akarray_dst:pdarray,
-                       akarray_weight:Union[None,pdarray] = None) -> None:
-        """
-        Populates the graph object with edges as defined by the akarrays. Uses an Arkouda-based
-        reading version. 
-
-        Returns
-        -------
-        None
-        """
-        cmd = "addEdgesFrom"
-
-        ### Edge dedupping and delooping.
-        # 1. Initialize the edge arrays.
-        src = akarray_src
-        dst = akarray_dst
-
-        # 1a. Initialize the weights array, if applicable.
-        wgt = ak.array([1.0])
-        if isinstance(akarray_weight,pdarray):
-            wgt = akarray_weight
-            self.weighted = 1
-
-        # 2. Sort the edges and remove duplicates.
-        gb_edges = ak.GroupBy([src,dst])
-        src = gb_edges.unique_keys[0]
-        dst = gb_edges.unique_keys[1]
-        if self.weighted == 1:
-            wgt = gb_edges.aggregate(wgt, "sum")[1]
-
-        ### Vertex remapping.
-        # 1. Extract the unique vertex names of the graph.
-        all_vertices = ak.concatenate([src,dst])
-        gb_vertices = ak.GroupBy(all_vertices)
-        vmap = gb_vertices.unique_keys
-
-        # 2. Create evenly spaced array within the range of the size of unique keys and broadcast
-        #    the values of the new range to the original vertices.
-        new_vertex_range = ak.arange(gb_vertices.unique_keys.size)
-        all_vertices = gb_vertices.broadcast(new_vertex_range)
-        src = all_vertices[0:src.size]
-        dst = all_vertices[src.size:]
-
-        ### Create vertex index arrays.
-        # 1. Build the out-neighbors of the adjacency lists for each vertex.
-        gb_src = ak.GroupBy(src, assume_sorted = True)
-        gb_src_indices, gb_src_neighbors = gb_src.count()
-        neis = ak.full(gb_vertices.unique_keys.size, 0, dtype=ak.int64)
-        neis[gb_src_indices] = gb_src_neighbors
-
-        # 2. Run a prefix (cumulative) sum on neis to get the starting indices for each vertex.
-        segs = ak.cumsum(neis)
-        first_seg = ak.array([0])
-        segs = ak.concatenate([first_seg, segs])
-
-        # 3. Extract the number of vertices, edges, and weighted flag from the graph.
-        self.n_vertices = int(vmap.size)
-        self.n_edges = int(src.size)
-
-        ### Create the reversed arrays for in-neighbor calculations.
-        # 1. Reverse the edges and sort them.
-        gb_edges_reversed = ak.GroupBy([dst, src])
-        src_reversed = gb_edges_reversed.unique_keys[0]
-        dst_reversed = gb_edges_reversed.unique_keys[1]
-
-        # 2. Build the in-neighbors of the adjacency lists for each vertex.
-        gb_src_reversed = ak.GroupBy(src_reversed, assume_sorted = True)
-        gb_src_indices_reversed, gb_src_neighbors_reversed = gb_src_reversed.count()
-        neis_reversed = ak.full(gb_vertices.unique_keys.size, 0, dtype=ak.int64)
-        neis_reversed[gb_src_indices_reversed] = gb_src_neighbors_reversed
-
-        # 3. Run a prefix (cumulative) sum on neis to get the starting indices for each vertex.
-        segs_reversed = ak.cumsum(neis_reversed)
-        first_seg_reversed = ak.array([0])
-        segs_reversed = ak.concatenate([first_seg_reversed, segs_reversed])
-
-        ### Store everything in a graph object in the Chapel server.
-        # 1. Store data into an Graph object in the Chapel server.
-        args = { "AkArraySrc" : src,
-                 "AkArrayDst" : dst,
-                 "AkArraySeg" : segs,
-                 "AkArraySrcR" : src_reversed,
-                 "AkArrayDstR" : dst_reversed,
-                 "AkArraySegR" : segs_reversed,
-                 "AkArrayWeight" : wgt,
-                 "AkArrayVmap" : vmap,
-                 "Directed": bool(self.directed),
-                 "Weighted" : bool(self.weighted),
-                 "NumVertices" : self.n_vertices,
-                 "NumEdges" : self.n_edges }
-
-        rep_msg = generic_msg(cmd=cmd, args=args)
-        oriname = rep_msg
-        self.name = oriname.strip()
-
